backend/requirements.txt
pika
Flask-SQLAlchemy==2.4.0
psycopg2-binary==2.8.3
urllib.parse
